Replace suffix cuda by suffix gpu in directory and file names
