+++ examples/gpu/ta_dense_gpu.cpp
@@ -24,8 +24,8 @@
 // clang-format off
 
 #include <tiledarray.h>
-#include <TiledArray/cuda/btas_um_tensor.h>
-#include "TiledArray/cuda/cpu_cuda_vector.h"
+#include <TiledArray/gpu/btas_um_tensor.h>
+#include "TiledArray/gpu/cpu_gpu_vector.h"
 #include <TiledArray/external/btas.h>
 // clang-format on
