+++ src/TiledArray/expressions/expr.h
@@ -41,7 +41,7 @@
 #include "TiledArray/tile_interface/trace.h"
 #include "expr_engine.h"
 #ifdef TILEDARRAY_HAS_CUDA
-#include <TiledArray/cuda/cuda_task_fn.h>
+#include <TiledArray/gpu/gpu_task_fn.h>
 #include <TiledArray/external/cuda.h>
 #endif
