+++ src/TiledArray/gpu/kernel/mult_kernel.cu
@@ -21,8 +21,8 @@
  *
  */
 
-#include <TiledArray/cuda/kernel/mult_kernel.h>
-#include <TiledArray/cuda/kernel/mult_kernel_impl.h>
+#include <TiledArray/gpu/kernel/mult_kernel.h>
+#include <TiledArray/gpu/kernel/mult_kernel_impl.h>
 
 
 #ifdef TILEDARRAY_HAS_CUDA
