+++ src/TiledArray/tiledarray.cpp
@@ -7,7 +7,7 @@
 #include <madness/world/safempi.h>
 
 #ifdef TILEDARRAY_HAS_CUDA
-#include <TiledArray/cuda/cublas.h>
+#include <TiledArray/gpu/cublas.h>
 #include <TiledArray/external/cuda.h>
 #include <librett.h>
 #endif
